Refactor Public Inputs and add them to the transcript (#134)
* Modify `construct_dense_pi_vec`

Update `construct_dense_pi_vec` so it returns the evaluation
representation of the PI polynomial over the used domain. This means
that the result vector will always have a power of two length.

* Add PI to the transcript

Co-authored-by: Joshua Fitzgerald <joshuabfitzgerald@gmail.com>

* Modify `construct_dense_pi_vec`

Update `construct_dense_pi_vec` so it returns the evaluation
representation of the PI polynomial over the used domain. This means
that the result vector will always have a power of two length.

* Add PI to the transcript

Co-authored-by: Joshua Fitzgerald <joshuabfitzgerald@gmail.com>

* New public inputs implementation

* Change `PI` implementation

The new implmenentation uses BTreeMap to ensure no duplicates are
introduced and to easily get an ordered representation. This is
particualrly helpuful to get a unique representation to add to the
transcript.

* New error type `InvalidPublicInputValue`

* Remove leftover

* Add non-zero public inputs to the transcript

* Adapt to new `PI` struct

* Fix intradoc links & lint checks

* Add serialized `PI` to the transcript.

Public Inputs are serialized and added to the transcript under the label
"pi". The struct PI is responsible of guaranteeing a unique
representation of each possible public input vector.

* Remove `Deref` from `PI`

* Replace `PrimeField` with `FftField`.

* Add basic tests

* Stray comments

* Use `derivative::Derivative` instead of `derive`

* Drop unnecessary comment

* Rename `PI` to `PublicInputs`

* Fix `extend`

* Fix intradoc links

Co-authored-by: Joshua Fitzgerald <joshuabfitzgerald@gmail.com>